Lowercase redirect URLs to fix #4391 and handle nested resources.
First issue is that the new static.wixtoolset.org site is case sensitive
and all URLs there are lowercase so ensure the download handler does
that to fix #4391. At the same time noticed, the split function was
being too greedy and separating the first part of the path from the
rest of the filename. That would cause any download file in a subfolder
on the static.wixtoolset.org to be incorrectly addressed.