Fix assignment agreement GitHub login
An update to TinyWebStack added additional meaning to the "Output"
property. Namely it defined which renderer to use to write the output
back to the response. Since there was no renderer for GithubUser, nor
was there ever intended to be one, TinyWebStack believed the handler
had no work to do for this client and returned HTTP 406.

Since the GithubUser was only needed in a local scope, this commit
removed the Output property and everything went back to working.