Update chain.md
All the links on this page points to the wrong page. 

If we take MsiPackage as an example the link directs you to: https://wixtoolset.org/docs/v3/xsd/wix/wix/msipackage/ when it should be https://wixtoolset.org/docs/v3/xsd/wix/msipackage/ . An additonal /wix is added to all the paths which will make the links faulty. The only link that worked was the one to the Wix schema. However, it did not go to the wix schema but instead the wix element page.